tagging solo realizar el tag, ahora se le especifica la rama. Por otro lado la infrastructura realiza las cosas mediate make

+++ scripts/infrastructure.sh
@@ -6,19 +6,22 @@ DATABASE_NAME="${2}"
 DATABASE_USER="${3}"
 DATABASE_PASSWORD="${4}"
 PATH_RELEASE="${5}"
-URL_IMAGES_TARGZ="${6}"
+CLEAN_UP="${6}"
 
-# Configuracion DB
-mv $PATH_RELEASE/includes/configure_example.php $PATH_RELEASE/includes/configure.php ; \
-sed -i -- "s/{DB_SERVER}/${DATABASE_HOST}/g" $PATH_RELEASE/includes/configure.php && \
-sed -i -- "s/{DB_USERNAME}/${DATABASE_USER}/g" $PATH_RELEASE/includes/configure.php && \
-sed -i -- "s/{DB_PASSWORD}/${DATABASE_PASSWORD}/g" $PATH_RELEASE/includes/configure.php && \
-sed -i -- "s/{DB_DATABASE}/${DATABASE_NAME}/g" $PATH_RELEASE/includes/configure.php ; \
-
-# Descarga de imagenes
-if [ $URL_IMAGES_TARGZ != "" ]; then
-	cd $PATH_RELEASE/images
-	wget -O images.tar.gz $URL_IMAGES_TARGZ >/dev/null 2>&1
-	tar zxf images.tar.gz
-	rm images.tar.gz
+# Reset
+if [ "$CLEAN_UP" = "true" ]; then 
+	# Descarga de imagenes
+	make -f $PATH_RELEASE/Makefile images
 fi
+
+# Comprobamos si no tiene configure
+if [[ -f  $PATH_RELEASE/includes/configure_example.php ]] && [[ ! -f  $PATH_RELEASE/includes/configure.php ]]; then
+	mv $PATH_RELEASE/includes/configure_example.php $PATH_RELEASE/includes/configure.php ; \
+	sed -i -- "s/{DB_SERVER}/${DATABASE_HOST}/g" $PATH_RELEASE/includes/configure.php && \
+	sed -i -- "s/{DB_USERNAME}/${DATABASE_USER}/g" $PATH_RELEASE/includes/configure.php && \
+	sed -i -- "s/{DB_PASSWORD}/${DATABASE_PASSWORD}/g" $PATH_RELEASE/includes/configure.php && \
+	sed -i -- "s/{DB_DATABASE}/${DATABASE_NAME}/g" $PATH_RELEASE/includes/configure.php ; \
+fi
+
+# Permisos
+make -f $PATH_RELEASE/Makefile permissions
